+++ game_patch/misc/misc.h
@@ -4,3 +4,6 @@ void MiscInit();
 void MiscAfterFullGameInit();
 void SetPlaySoundEventsVolumeScale(float volume_scale);
 void SetJumpToMultiServerList(bool jump);
+
+constexpr int old_obj_limit = 1024;
+constexpr int obj_limit = 65536;